backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.12
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.22
langchain-pinecone==0.2.8
langchain-tests==0.3.20
langchain-text-splitters==0.3.8
langsmith==0.3.45
markdown-it-py==3.0.0
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.3.0
openai==1.86.0
orjson==3.10.18
packaging==24.2
pinecone==7.0.2
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
propcache==0.3.2
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pypdf==5.6.0
pytest==8.4.0
pytest-asyncio==0.26.0
pytest-benchmark==5.1.0
pytest-codspeed==3.2.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
syrupy==4.9.1
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
vcrpy==7.0.0
wrapt==1.17.2
yarl==1.20.1
zstandard==0.23.0
cachetools==5.5.2
redis==5.2.1
google-genai==1.14.0
google-cloud-storage==2.19.0
google-cloud-discoveryengine==0.13.5
PyMySQL
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart==0.0.20
python-docx==1.2.0
pdfplumber==0.11.1